requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
lxml==4.8.0
Pillow==9.0.1
pyttsx3==2.90
requests==2.27.1
SpeechRecognition==3.8.1
soupsieve==2.3.1
urllib3==1.26.8
youtube-dl==2021.4.26
PyAudio==0.2.11